Revert "Initial cut at a failed game concept - Smash or Crash :) Turns out the OMDB API isn't capable of such a thing."
This reverts commit c3ef0a0c727d62bcfa15b2abc53c3c4573c9b9e7.
